+++ src/helics/core/flagOperations.hpp
@@ -43,7 +43,7 @@ constexpr uint16_t slow_responding_flag =
 template<class FlagContainer, class FlagIndex>
 inline void setActionFlag(FlagContainer& M, FlagIndex flag)
 {
-    M.flags |= (static_cast<decltype(M.flags)>(1) << (static_cast<uint16_t>((flag))));
+    M.flags |= (static_cast<decltype(M.flags)>((1U) << (static_cast<uint16_t>((flag)))));
 }
 
 /** check a flag value on a specified index*/
@@ -51,7 +51,7 @@ inline void setActionFlag(FlagContainer& M, FlagIndex flag)
 template<class FlagIndex>
 inline bool checkActionFlag(uint16_t flags, FlagIndex flag)
 {
-    return ((flags & (static_cast<uint16_t>(1) << (static_cast<uint16_t>((flag))))) != 0U);
+    return ((flags & (static_cast<uint16_t>((1U) << (static_cast<uint16_t>((flag)))))) != 0U);
 }
 
 /** template function to check a flag in an object containing a flags field*/
@@ -59,20 +59,21 @@ template<class FlagContainer, class FlagIndex>
 inline bool checkActionFlag(const FlagContainer& M, FlagIndex flag)
 {
     return (
-        (M.flags & (static_cast<decltype(M.flags)>(1) << (static_cast<uint16_t>((flag))))) != 0U);
+        (M.flags & (static_cast<decltype(M.flags)>((1U) << (static_cast<uint16_t>((flag)))))) !=
+        0U);
 }
 
 /** template function to clear a flag in an object containing a flags field*/
 template<class FlagContainer, class FlagIndex>
 inline void clearActionFlag(FlagContainer& M, FlagIndex flag)
 {
-    M.flags &= ~(static_cast<decltype(M.flags)>(1) << (static_cast<uint16_t>((flag))));
+    M.flags &= static_cast<decltype(M.flags)>(~((1U) << (static_cast<uint16_t>((flag)))));
 }
 
 /** helper function to facilitate make a flag variable*/
 inline constexpr uint16_t make_flags(unsigned int flag)
 {
-    return static_cast<uint16_t>(1) << (flag);
+    return static_cast<uint16_t>(1U << (flag));
 }
 
 /** helper function to facilitate make a flag variable out of two flags*/