+++ src/helics/shared_api_library/MessageFederate.h
@@ -165,7 +165,7 @@ HELICS_EXPORT void helicsEndpointSendEventRaw(
 
 /**
  * Send a message object from a specific endpoint.
- *
+ * @deprecated Use helicsEndpointSendMessageObject instead.
  * @param endpoint The endpoint to send the data from.
  * @param message The actual message to send.
  * @forcpponly
@@ -173,17 +173,30 @@ HELICS_EXPORT void helicsEndpointSendEventRaw(
  * @endforcpponly
  */
 
-HELICS_EXPORT void helicsEndpointSendMessage(helics_endpoint endpoint, helics_message* message, helics_error* err);
+HELICS_DEPRECATED_EXPORT void helicsEndpointSendMessage(helics_endpoint endpoint, helics_message* message, helics_error* err);
+
 /**
  * Send a message object from a specific endpoint.
  *
  * @param endpoint The endpoint to send the data from.
- * @param message The actual message to send.
+ * @param message The actual message to send which will be copied.
  * @forcpponly
  * @param[in,out] err A pointer to an error object for catching errors.
  * @endforcpponly
  */
 HELICS_EXPORT void helicsEndpointSendMessageObject(helics_endpoint endpoint, helics_message_object message, helics_error* err);
+
+/**
+ * Send a message object from a specific endpoint, the message will not be copied and the message object will no longer be valid
+ * after the call.
+ *
+ * @param endpoint The endpoint to send the data from.
+ * @param message The actual message to send which will be copied.
+ * @forcpponly
+ * @param[in,out] err A pointer to an error object for catching errors.
+ * @endforcpponly
+ */
+HELICS_EXPORT void helicsEndpointSendMessageObjectZeroCopy(helics_endpoint endpoint, helics_message_object message, helics_error* err);
 
 /**
  * Subscribe an endpoint to a publication.
@@ -231,11 +244,14 @@ HELICS_EXPORT int helicsEndpointPendingMessages(helics_endpoint endpoint);
 /**
  * Receive a packet from a particular endpoint.
  *
+ * @deprecated This function is deprecated and will be removed in Helics 3.0.
+ *             Use helicsEndpointGetMessageObject instead.
+ *
  * @param[in] endpoint The identifier for the endpoint.
  *
  * @return A message object.
  */
-HELICS_EXPORT helics_message helicsEndpointGetMessage(helics_endpoint endpoint);
+HELICS_DEPRECATED_EXPORT helics_message helicsEndpointGetMessage(helics_endpoint endpoint);
 
 /**
  * Receive a packet from a particular endpoint.
@@ -249,13 +265,16 @@ HELICS_EXPORT helics_message_object helicsEndpointGetMessageObject(helics_endpoi
 /**
  * Receive a communication message for any endpoint in the federate.
  *
+ * @deprecated This function is deprecated and will be removed in Helics 3.0.
+ *             Use helicsFederateGetMessageObject instead.
+ *
  * @details The return order will be in order of endpoint creation.
  *          So all messages that are available for the first endpoint, then all for the second, and so on.
  *          Within a single endpoint, the messages are ordered by time, then source_id, then order of arrival.
  *
  * @return A unique_ptr to a Message object containing the message data.
  */
-HELICS_EXPORT helics_message helicsFederateGetMessage(helics_federate fed);
+HELICS_DEPRECATED_EXPORT helics_message helicsFederateGetMessage(helics_federate fed);
 
 /**
  * Receive a communication message for any endpoint in the federate.
@@ -289,9 +308,13 @@ HELICS_EXPORT void helicsFederateClearMessages(helics_federate fed);
 /**
  * Clear all message from an endpoint.
  *
+ * @deprecated This function does nothing and will be removed. 
+ *             Use helicsFederateClearMessages to free all messages,
+ *             or helicsMessageFree to clear an individual message.
+ * 
  * @param endpoint The endpoint object to operate on.
  */
-HELICS_EXPORT void helicsEndpointClearMessages(helics_endpoint endpoint);
+HELICS_DEPRECATED_EXPORT void helicsEndpointClearMessages(helics_endpoint endpoint);
 
 /**
  * Get the type specified for an endpoint.