+++ tests/helics/shared_library/test-message-federate_cpp.cpp
@@ -89,10 +89,10 @@ TEST_P(mfed_type_tests, message_federate_send_receive)
 
     auto M = epid2.getMessage();
     // BOOST_REQUIRE (M);
-    ASSERT_EQ(M.length, 500);
-    EXPECT_NE(M.data, nullptr);
-    if (M.data != nullptr) {
-        EXPECT_EQ(M.data[245], 'a');
+    ASSERT_EQ(M.size(), 500);
+    EXPECT_NE(M.data(), nullptr);
+    if (M.data() != nullptr) {
+        EXPECT_EQ(M.c_str()[245], 'a');
     }
     mFed1->finalize();